영업기회 리스트 출력 전체 코드 및 레이아웃/css 코드 추가/수정

+++ src/main/java/kr/co/bizcore/v1/domain/Sopp.java
@@ -1,5 +1,6 @@
 package kr.co.bizcore.v1.domain;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
@@ -34,7 +35,7 @@ public class Sopp extends SimpleSopp{
     private int cntrctMth;
     private String soppTitle;
     private String soppDesc;
-    private int soppTargetAmt;
+    private BigDecimal soppTargetAmt;
     private String soppTargetDate;
     private String maintenance_S;
     private String maintenance_E;
